+++ static/js/notifications.js
@@ -1,0 +1,3 @@
+$(function () {
+    $("#tz").val(Intl.DateTimeFormat().resolvedOptions().timeZone);
+});