+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Django 3.2.2
 - Improve the handling of unknown email addresses in the Sign In form
 - Add support for "... is UP" SMS notifications
+- Add an option for weekly reports (in addition to monthly)
 
 ## v1.20.0 - 2020-04-22